% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{create_directory}
\alias{create_directory}
\title{Create directory}
\usage{
create_directory(path)
}
\arguments{
\item{path}{path to create directory in}
}
\description{
Test for existence of directory; if directory does not exist,
attempt to recursively create directories to path
}
