% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy-files.R
\name{import_template_files}
\alias{import_template_files}
\title{Import template files}
\usage{
import_template_files(target_path = NULL)
}
\arguments{
\item{target_path}{directory to import template files to}
}
\description{
Import template files
}
