% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{normalise_path}
\alias{normalise_path}
\title{Convert path}
\usage{
normalise_path(path)
}
\arguments{
\item{path}{path to normalise}
}
\description{
Convert paths into the canonical form for the platform;
on Windows this is "C:/Parent/Child"
A simple wrapper for the normalizePath() function
}
