% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-binding.R
\name{site_skeleton}
\alias{site_skeleton}
\title{Setting out site skeleton}
\usage{
site_skeleton(path)
}
\arguments{
\item{path}{the path where the website goes.}
}
\description{
Creating the skeleton of an Rmarkdown website
}
